+++ samples/rad/main.cpp
@@ -219,7 +219,6 @@ int main(int argc, char* argv[]) {
   auto console = tcod::Console{80, 50};
   auto tileset = tcod::load_tilesheet("data/fonts/terminal8x8_gs_tc.png", {32, 8}, tcod::CHARMAP_TCOD);
   TCOD_ContextParams params{};
-  params.tcod_version = TCOD_COMPILEDVERSION;
   params.console = console.get();
   params.tileset = tileset.get();
   params.vsync = true;