+++ src/libtcod/console_init.c
@@ -61,7 +61,7 @@ TCOD_Error TCOD_console_init_root_(
   strncpy(TCOD_ctx.window_title, title ? title : "", sizeof(TCOD_ctx.window_title) - 1);
   TCOD_ctx.fullscreen = fullscreen;
   struct TCOD_ContextParams params = {
-      .tcod_version = TCOD_COMPILEDVERSION,
+      .tcod_version = 0,
       .window_x = SDL_WINDOWPOS_UNDEFINED,
       .window_y = SDL_WINDOWPOS_UNDEFINED,
       .columns = w,